Marconi.Api.UtxoIndexersQuery

Documentation

bootstrap

Arguments

:: TargetAddresses
user provided target addresses

-> IO DBQueryEnv
returns Query runtime environment

Bootstraps the utxo query environment. The module is responsible for accessing SQLite for quries. The main issue we try to avoid here is mixing inserts and quries in SQLite to avoid locking the database

findByCardanoAddress

Arguments

:: DBQueryEnv
Query run time environment

-> AddressAny
Cardano address to query

-> IO [UtxoRow]

Query utxos by Cardano Address To Cardano error may occure

findByAddress

Arguments

:: DBQueryEnv
Query run time environment

-> Text
Bech32 Address

-> IO (Either QueryExceptions UtxoTxOutReport)
To Plutus address conversion error may occure

Retrieve a Set of TxOutRefs associated with the given Cardano Era address We return an empty Set if no address is found

findAll

Arguments

:: DBQueryEnv
Query run time environment

-> IO [UtxoTxOutReport]
set of corresponding TxOutRefs

finds reports for all user-provided addresses. TODO consider sqlite streaming, https://hackage.haskell.org/package/sqlite-simple-0.4.18.2/docs/Database-SQLite-Simple.html#g:14

reportQueryAddresses :: DBQueryEnv -> IO [Address ShelleyAddr]

report target addresses Used by JSON-RPC

data UtxoRow

Constructors

UtxoRow

Fields

_utxoRowUtxo :: Utxo

_utxoRowSlotNo :: !SlotNo

_utxoRowBlockNo :: !BlockNo

type UtxoIndex = SqliteIndex UtxoEvent Notification AddressAny Result

reportQueryCardanoAddresses :: DBQueryEnv -> Text

reportBech32Addresses :: DBQueryEnv -> [Text]

withQueryAction

Arguments

:: DBQueryEnv
Query run time environment

-> AddressAny
Cardano address to query

-> IO [UtxoRow]

Execute the query function We must stop the utxo inserts before doing the query

writeTMVar :: TMVar a -> a -> STM ()

Non-blocking write of a new value to a TMVar Puts if empty. Replaces if populated.