Marconi.Bootstrap

Description

This module bootstraps the mamba JSON RPC server, it acts as a glue conntecting the JSON-RPC, HttpServer, marconiIndexer, and marconi cache

Documentation

bootstrapJsonRpc :: Maybe RpcPortNumber -> TargetAddresses -> IO JsonRpcEnv

Bootstraps the JSON-RPC http server with appropriate settings and marconi cache this is just a wrapper for the bootstrapHttp in json-rpc package

bootstrapHttp :: JsonRpcEnv -> IO ()

bootstrapUtxoIndexers :: CliArgs -> JsonRpcEnv -> IO ()

marconi cardano blockchain indexer

targetAddressParser

Arguments

:: String
contains white spece delimeted lis of addresses

-> TargetAddresses
a non empty list of valid addresses

parses a white space separated address list Note, duplicate addresses are rmoved

fromJustWithError :: Show e => Either e a -> a

Exit program with error Note, if the targetAddress parser fails, or is empty, there is nothing to do for the hotStore. In such case we should fail fast